+++ lib/src/refs/provider/multi_provider.dart
@@ -16,6 +16,9 @@ class MultiProvider extends RefWidget<void> {
   final List<Provider> providers;
 
   @override
+  Function? get create => null;
+
+  @override
   RefState<void, MultiProvider> createState() => MultiProviderState();
 }